Added a bash script that will automatically configure guzzoni.apple.com in dnsmasq to point to a specific domain (in this case, my dynamic ip tracker).
The script runs as a elevated cronjob every half an hour. This is not a great fix, but it
at least eliminates out one of the two devices I have to reset with new ip addresses as the
siriproxy server now knows to fetch it's own ip address.

Perhaps I could tie this in to a script that will also change the index.html on my
webserver to the current IP address so I can look it up easily from my phone.